Integration UI for sync contact v2
