--- lib/data/datasource/contact/phonebook_datasource_v2.dart
@@ -1,5 +1,0 @@
-import 'package:fluffychat/domain/model/contact/contact_new.dart';
-
-abstract class PhonebookContactDatasourceV2 {
-  Future<List<Contact>> fetchContacts();
-}
