+++ lib/data/datasource_impl/contact/phonebook_contact_datasource_impl.dart
@@ -1,83 +1,114 @@
-import 'package:fluffychat/data/datasource/phonebook_datasouce.dart';
 import 'package:fluffychat/domain/model/contact/contact.dart';
 import 'package:fluffychat/utils/string_extension.dart';
-import 'package:flutter_contacts/flutter_contacts.dart' hide Contact;
+import 'package:flutter_contacts/flutter_contacts.dart' as flutter_contact;
+import 'package:fluffychat/data/datasource/contact/phonebook_datasource.dart';
 
 class PhonebookContactDatasourceImpl implements PhonebookContactDatasource {
   @override
   Future<List<Contact>> fetchContacts() async {
     final phonebookContacts =
-        await FlutterContacts.getContacts(withProperties: true);
+        await flutter_contact.FlutterContacts.getContacts(withProperties: true);
 
-    final listPhoneContacts = phonebookContacts
-        .expand(
-          (contact) => contact.phones.map(
-            (phone) => Contact(
-              phoneNumber: phone.number,
-              displayName: contact.displayName,
-            ),
-          ),
-        )
-        .toList();
+    final contacts = mappingToContact(phonebookContacts);
+    return _sortContactsByDisplayName(contacts);
+  }
 
-    final listEmailContacts = phonebookContacts
-        .expand(
-          (contact) => contact.emails.map(
-            (email) => Contact(
-              email: email.address,
-              displayName: contact.displayName,
-            ),
-          ),
-        )
-        .toList();
+  List<Contact> mappingToContact(
+    List<flutter_contact.Contact> phoneBookContacts,
+  ) {
+    final listAllContacts = phoneBookContacts.map((contact) {
+      final phoneNumbers = contact.phones
+          .map((phone) => PhoneNumber(number: phone.number))
+          .toList();
+      final emails =
+          contact.emails.map((email) => Email(address: email.address)).toList();
 
-    final listAllContacts = [
-      ..._removeDuplicatedPhoneNumbers(listPhoneContacts),
-      ...listEmailContacts,
-    ];
+      return Contact(
+        id: contact.id,
+        displayName: contact.displayName,
+        phoneNumbers: _removeDuplicatedPhoneNumbers(phoneNumbers).toSet(),
+        emails: _removeDuplicatedEmails(emails).toSet(),
+      );
+    }).toList();
 
     return listAllContacts;
   }
 
-  List<Contact> _removeDuplicatedPhoneNumbers(List<Contact> listContacts) {
+  List<Contact> _sortContactsByDisplayName(List<Contact> contacts) {
+    contacts.sort(
+      (pre, next) => (pre.displayName ?? '').compareTo(next.displayName ?? ''),
+    );
+    return contacts;
+  }
+
+  List<PhoneNumber> _removeDuplicatedPhoneNumbers(
+    List<PhoneNumber> phoneNumbers,
+  ) {
     final listVisitedPhoneNumbers = <String>[];
-    final listFilteredContacts = <Contact>[];
+    final filteredPhoneNumbers = <PhoneNumber>[];
 
-    final listContactHasPhoneNumber =
-        listContacts.where((contact) => contact.phoneNumber != null).toList();
-
-    for (final contact in listContactHasPhoneNumber) {
-      final phoneNumber = contact.phoneNumber;
-      final normalizedPhoneNumber = phoneNumber!.normalizePhoneNumber();
+    for (final contact in phoneNumbers) {
+      final phoneNumber = contact.number;
+      final normalizedPhoneNumber = phoneNumber.normalizePhoneNumber();
 
       if (listVisitedPhoneNumbers.contains(normalizedPhoneNumber)) {
-        final hasSameFilteredContact = _hasSameFilteredContact(
-          listFilteredContacts,
-          contact.displayName ?? '',
+        final hasSameFilteredContact = _hasSameFilteredPhoneNumber(
+          filteredPhoneNumbers,
           normalizedPhoneNumber,
         );
         if (!hasSameFilteredContact) {
-          listFilteredContacts.add(contact);
+          filteredPhoneNumbers.add(contact);
         }
       } else {
         listVisitedPhoneNumbers.add(normalizedPhoneNumber);
-        listFilteredContacts.add(contact);
+        filteredPhoneNumbers.add(contact);
       }
     }
 
-    return listFilteredContacts;
+    return filteredPhoneNumbers;
   }
 
-  bool _hasSameFilteredContact(
-    List<Contact> listFilteredContacts,
-    String contactName,
+  bool _hasSameFilteredPhoneNumber(
+    List<PhoneNumber> filteredPhoneNumbers,
     String phoneNumberNormalized,
   ) {
-    return listFilteredContacts.any(
+    return filteredPhoneNumbers.any(
       (filteredContact) =>
-          filteredContact.displayName == contactName &&
-          filteredContact.phoneNumber?.normalizePhoneNumber() ==
-              phoneNumberNormalized,
+          filteredContact.number.normalizePhoneNumber() ==
+          phoneNumberNormalized,
+    );
+  }
+
+  List<Email> _removeDuplicatedEmails(List<Email> emails) {
+    final listVisitedEmails = <String>[];
+    final filteredEmails = <Email>[];
+
+    for (final email in emails) {
+      final emailAddress = email.address;
+
+      if (listVisitedEmails.contains(emailAddress)) {
+        final hasSameFilteredEmail = _hasSameFilteredEmail(
+          filteredEmails,
+          emailAddress,
+        );
+        if (!hasSameFilteredEmail) {
+          filteredEmails.add(email);
+        }
+      } else {
+        listVisitedEmails.add(emailAddress);
+        filteredEmails.add(email);
+      }
+    }
+
+    return filteredEmails;
+  }
+
+  bool _hasSameFilteredEmail(
+    List<Email> filteredEmails,
+    String emailAddress,
+  ) {
+    return filteredEmails.any(
+      (filteredEmail) => filteredEmail.address == emailAddress,
     );
   }
 }
