+++ lib/data/network/contact/lookup_api.dart
@@ -1,14 +1,18 @@
+import 'dart:io';
+
+import 'package:dio/dio.dart';
 import 'package:fluffychat/data/network/dio_client.dart';
 import 'package:fluffychat/data/network/identity_endpoint.dart';
 import 'package:fluffychat/di/global/get_it_initializer.dart';
 import 'package:fluffychat/di/global/network_di.dart';
+import 'package:fluffychat/domain/model/contact/federation_identity_register_response.dart';
 import 'package:fluffychat/domain/model/contact/hash_details_response.dart';
 import 'package:fluffychat/domain/model/contact/lookup_list_mxid_request.dart';
 import 'package:fluffychat/domain/model/contact/lookup_list_mxid_response.dart';
 
 class LookupAPI {
   final DioClient _client =
-      getIt.get<DioClient>(instanceName: NetworkDI.identityDioClientName);
+      getIt.get<DioClient>(instanceName: NetworkDI.federationDioClientName);
 
   Future<HashDetailsResponse> getHashDetails() async {
     final path = IdentityEndpoint.hashDetailsServicePath
@@ -26,5 +30,26 @@ class LookupAPI {
       data: request.toJson(),
     );
     return LookupListMxidResponse.fromJson(response);
+  }
+
+  Future<FederationIdentityRegisterResponse> federationIdentityRegister({
+    required dynamic data,
+  }) async {
+    final path = IdentityEndpoint.accountRegisterServicePath
+        .generateMatrixIdentityEndpoint();
+
+    final headers = {
+      HttpHeaders.acceptHeader: NetworkDI.acceptHeaderDefault,
+      HttpHeaders.contentTypeHeader: NetworkDI.contentTypeHeaderDefault,
+    };
+
+    final response = await _client.postToGetBody(
+      path,
+      data: data,
+      options: Options(
+        headers: headers,
+      ),
+    );
+    return FederationIdentityRegisterResponse.fromJson(response);
   }
 }
