+++ lib/data/network/interceptor/federation_authorization_interceptor.dart
@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:dio/dio.dart';
 import 'package:matrix/matrix.dart';
 
-class AuthorizationInterceptor extends InterceptorsWrapper {
-  AuthorizationInterceptor();
+class FederationAuthorizationInterceptor extends InterceptorsWrapper {
+  FederationAuthorizationInterceptor();
 
   String? _accessToken;
 
@@ -12,12 +12,12 @@ class AuthorizationInterceptor extends InterceptorsWrapper {
     _accessToken = accessToken;
   }
 
-  String? get getAccessToken => _accessToken;
-
   @override
   void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
     options.headers[HttpHeaders.authorizationHeader] = _bearerToken;
-    Logs().d('AuthorizationInterceptor::onRequest:accessToken: $_bearerToken');
+    Logs().d(
+      'FederationAuthorizationInterceptor::onRequest:accessToken: $_bearerToken',
+    );
     super.onRequest(options, handler);
   }
