+++ lib/domain/model/extensions/string_extension.dart
@@ -8,4 +8,9 @@ extension StringExtension on String {
   String get displayMentioned {
     return '@[$this]';
   }
+
+  String get convertToHttps {
+    final domain = split(':').first;
+    return 'https://$domain/';
+  }
 }
