--- lib/domain/repository/phonebook_contact_repository_v2.dart
@@ -1,5 +1,0 @@
-import 'package:fluffychat/domain/model/contact/contact_new.dart';
-
-abstract class PhonebookContactRepositoryV2 {
-  Future<List<Contact>> fetchContacts();
-}
