+++ lib/modules/federation_identity_lookup/data/network/federation_identity_endpoint.dart
@@ -20,7 +20,7 @@ class FederationIdentityEndpoint {
   );
 
   static final ServicePath registerServicePath = ServicePath(
-    '/register',
+    '/account/register',
   );
 }
