+++ lib/modules/federation_identity_request_token/data/network/federation_identity_request_token_api.dart
@@ -13,7 +13,8 @@ class FederationIdentityRequestTokenApi {
     final path =
         FederationIdentityRequestTokenEndpoint.requestTokenServicePath(mxid)
             .generateFederationIdentityRequestTokenEndpoint();
-    final response = await client.postToGetBody(path);
-    return response;
+    final response = await client.postToGetBody(path, data: {});
+
+    return FederationTokenInformation.fromJson(response);
   }
 }
