+++ lib/pages/new_private_chat/new_private_chat.dart
@@ -88,7 +88,10 @@ class NewPrivateChatController extends State<NewPrivateChat>
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) async {
-    await handleDidChangeAppLifecycleState(state);
+    await handleDidChangeAppLifecycleState(
+      state,
+      client: Matrix.of(context).client,
+    );
   }
 
   @override
