+++ lib/presentation/mixins/contacts_view_controller_mixin.dart
@@ -4,12 +4,11 @@ import 'package:fluffychat/app_state/failure.dart';
 import 'package:fluffychat/app_state/success.dart';
 import 'package:fluffychat/di/global/get_it_initializer.dart';
 import 'package:fluffychat/domain/app_state/contact/get_contacts_state.dart';
-import 'package:fluffychat/domain/app_state/contact/get_phonebook_contact_state_v2.dart';
-import 'package:fluffychat/domain/app_state/contact/get_phonebook_contacts_state.dart';
+import 'package:fluffychat/domain/app_state/contact/get_phonebook_contact_state.dart';
 import 'package:fluffychat/domain/app_state/search/search_state.dart';
 import 'package:fluffychat/domain/contact_manager/contacts_manager.dart';
 import 'package:fluffychat/domain/model/contact/contact_type.dart';
-import 'package:fluffychat/domain/model/extensions/contact/contacts_extension.dart';
+import 'package:fluffychat/domain/model/extensions/contact/contact_extension.dart';
 import 'package:fluffychat/domain/usecase/search/search_recent_chat_interactor.dart';
 import 'package:fluffychat/presentation/enum/contacts/warning_contacts_banner_enum.dart';
 import 'package:fluffychat/presentation/extensions/contact/presentation_contact_extension.dart';
@@ -25,6 +24,7 @@ import 'package:fluffychat/utils/extension/presentation_search_extension.dart';
 import 'package:fluffychat/utils/permission_dialog.dart';
 import 'package:fluffychat/utils/permission_service.dart';
 import 'package:fluffychat/utils/platform_infos.dart';
+import 'package:fluffychat/widgets/matrix.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:matrix/matrix.dart';
@@ -64,7 +64,7 @@ mixin class ContactsViewControllerMixin {
 
   final presentationPhonebookContactV2Notifier =
       ValueNotifierCustom<Either<Failure, Success>>(
-    const Right(GetPhonebookContactsV2Initial()),
+    const Right(GetPhonebookContactsInitial()),
   );
 
   final FocusNode searchFocusNode = FocusNode();
@@ -99,7 +99,9 @@ mixin class ContactsViewControllerMixin {
             onRefuseTap: _handleDenyPermissionDialog,
             onAcceptButton: () async {
               Navigator.of(dialogContext).pop();
-              await _handleRequestContactsPermission();
+              await _handleRequestContactsPermission(
+                client: Matrix.of(context).client,
+              );
             },
           );
         },
@@ -135,7 +137,10 @@ mixin class ContactsViewControllerMixin {
     }
   }
 
-  Future<void> handleDidChangeAppLifecycleState(AppLifecycleState state) async {
+  Future<void> handleDidChangeAppLifecycleState(
+    AppLifecycleState state, {
+    required Client client,
+  }) async {
     if (!PlatformInfos.isMobile) {
       return;
     }
@@ -164,7 +169,9 @@ mixin class ContactsViewControllerMixin {
           currentContactPermission.isGranted) {
         contactsPermissionStatus = currentContactPermission;
         warningBannerNotifier.value = WarningContactsBannerState.hide;
-        contactsManager.refreshPhonebookContacts();
+        contactsManager.refreshPhonebookContacts(
+          mxid: client.userID!,
+        );
         return;
       }
     }
@@ -203,6 +210,7 @@ mixin class ContactsViewControllerMixin {
       );
     });
     contactsManager.initialSynchronizeContacts(
+      mxid: client.userID!,
       isAvailableSupportPhonebookContacts: PlatformInfos.isMobile &&
           contactsPermissionStatus != null &&
           contactsPermissionStatus == PermissionStatus.granted,
@@ -229,7 +237,7 @@ mixin class ContactsViewControllerMixin {
           ),
         );
 
-    contactsManager.getPhonebookContactsV2Notifier().addListener(
+    contactsManager.getPhonebookContactsNotifier().addListener(
           () => _refreshAllContacts(
             context: context,
             client: client,
@@ -245,8 +253,7 @@ mixin class ContactsViewControllerMixin {
   }) {
     final keyword = _debouncer.value;
     _refreshContacts(keyword);
-    // _refreshPhoneBookContacts(keyword);
-    _refreshPhoneBookContactsV2(keyword);
+    _refreshPhoneBookContacts(keyword);
     _refreshRecentContacts(
       context: context,
       client: client,
@@ -328,66 +335,12 @@ mixin class ContactsViewControllerMixin {
     );
   }
 
-  // Future<void> _refreshPhoneBookContacts(String keyword) async {
-  //   if (presentationPhonebookContactNotifier.isDisposed) return;
-  //   presentationPhonebookContactNotifier.value =
-  //       contactsManager.getPhonebookContactsNotifier().value.fold(
-  //     (failure) {
-  //       if (failure is GetPhonebookContactsFailure) {
-  //         return _handleSearchExternalContact(
-  //           keyword,
-  //           otherResult: Left(
-  //             GetPresentationContactsFailure(
-  //               keyword: keyword,
-  //             ),
-  //           ),
-  //         );
-  //       }
-  //
-  //       if (failure is GetPhonebookContactsIsEmpty) {
-  //         return _handleSearchExternalContact(
-  //           keyword,
-  //           otherResult: Left(
-  //             GetPresentationContactsEmpty(
-  //               keyword: keyword,
-  //             ),
-  //           ),
-  //         );
-  //       }
-  //       return Left(failure);
-  //     },
-  //     (success) {
-  //       if (success is GetPhonebookContactsSuccess) {
-  //         final filteredContacts = success.contacts
-  //             .searchContacts(keyword)
-  //             .expand((contact) => contact.toPresentationContacts())
-  //             .toList();
-  //         if (filteredContacts.isEmpty) {
-  //           return Left(
-  //             GetPresentationContactsEmpty(
-  //               keyword: keyword,
-  //             ),
-  //           );
-  //         } else {
-  //           return Right(
-  //             GetPresentationContactsSuccess(
-  //               contacts: filteredContacts,
-  //               keyword: keyword,
-  //             ),
-  //           );
-  //         }
-  //       }
-  //       return Right(success);
-  //     },
-  //   );
-  // }
-
-  Future<void> _refreshPhoneBookContactsV2(String keyword) async {
+  Future<void> _refreshPhoneBookContacts(String keyword) async {
     if (presentationPhonebookContactV2Notifier.isDisposed) return;
     presentationPhonebookContactV2Notifier.value =
-        contactsManager.getPhonebookContactsV2Notifier().value.fold(
+        contactsManager.getPhonebookContactsNotifier().value.fold(
       (failure) {
-        if (failure is GetPhonebookContactsV2Failure) {
+        if (failure is GetPhonebookContactsFailure) {
           return _handleSearchExternalContact(
             keyword,
             otherResult: Left(
@@ -398,7 +351,7 @@ mixin class ContactsViewControllerMixin {
           );
         }
 
-        if (failure is GetPhonebookContactsV2IsEmpty) {
+        if (failure is GetPhonebookContactsIsEmpty) {
           return _handleSearchExternalContact(
             keyword,
             otherResult: Left(
@@ -411,7 +364,7 @@ mixin class ContactsViewControllerMixin {
         return Left(failure);
       },
       (success) {
-        if (success is GetPhonebookContactsV2Success) {
+        if (success is GetPhonebookContactsSuccess) {
           final filteredContacts = success.contacts
               .searchContacts(keyword)
               .expand((contact) => contact.toPresentationContacts())
@@ -556,11 +509,15 @@ mixin class ContactsViewControllerMixin {
     isSearchModeNotifier.value = false;
   }
 
-  Future<void> _handleRequestContactsPermission() async {
+  Future<void> _handleRequestContactsPermission({
+    required Client client,
+  }) async {
     final currentContactsPermissionStatus =
         await _permissionHandlerService.requestContactsPermissionActions();
     if (currentContactsPermissionStatus == PermissionStatus.granted) {
-      contactsManager.refreshPhonebookContacts();
+      contactsManager.refreshPhonebookContacts(
+        mxid: client.userID!,
+      );
       warningBannerNotifier.value = WarningContactsBannerState.hide;
     } else {
       contactsManager.updateNotShowWarningContactsDialogAgain = true;
