+++ test/data/datasource_impl/contact/phonebook_contact_datasource_impl_test.dart
@@ -11,23 +11,42 @@ void main() {
   const MethodChannel channel =
       MethodChannel('github.com/QuisApp/flutter_contacts');
 
-  group('[PhonebookContactDatasourceImpl]', () {
+  group('[PhonebookContactDatasourceV2Impl] test\n', () {
     final listAllContacts = [
       {
+        'id': 'id_1',
         'displayName': 'Alice',
         'phones': [
           {'number': '(212)555-6789'},
           {'number': '2125556789'},
         ],
+        'emails': [
+          {
+            'address': 'Alice@domain.com',
+          },
+          {
+            'address': 'Alice_1@domain.com',
+          },
+        ],
       },
       {
+        'id': 'id_2',
         'displayName': 'Bob',
         'phones': [
           {'number': '2124678190'},
           {'number': '(212)467-8190'},
         ],
+        'emails': [
+          {
+            'address': 'bob@domain.com',
+          },
+          {
+            'address': 'bob2@domain.com',
+          }
+        ],
       },
       {
+        'id': 'id_3',
         'displayName': 'Charlie',
         'phones': [
           {'number': '212 555-6789'},
@@ -35,6 +54,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_4',
         'displayName': 'David',
         'phones': [
           {'number': '2124678190'},
@@ -42,6 +62,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_5',
         'displayName': 'Eve',
         'phones': [
           {'number': '+1.123.456.7890'},
@@ -49,6 +70,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_6',
         'displayName': 'Frank',
         'phones': [
           {'number': '81234977890'},
@@ -56,6 +78,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_7',
         'displayName': 'Grace',
         'phones': [
           {'number': '+1 (800)-555-1234 ext. 123'},
@@ -63,6 +86,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_8',
         'displayName': 'Hank',
         'phones': [
           {'number': '18005879106234'},
@@ -70,6 +94,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_9',
         'displayName': 'Ivy',
         'phones': [
           {'number': '+1 (800)-555.1234'},
@@ -77,6 +102,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_10',
         'displayName': 'Karl',
         'phones': [
           {'number': '18005873456'},
@@ -84,6 +110,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_11',
         'displayName': 'Liam',
         'phones': [
           {'number': '(212) 555-6789'},
@@ -91,6 +118,7 @@ void main() {
         ],
       },
       {
+        'id': 'id_12',
         'displayName': 'Mia',
         'phones': [
           {'number': '2125556789'},
@@ -98,10 +126,20 @@ void main() {
         ],
       },
       {
+        'id': 'id_13',
         'displayName': 'Nina',
         'emails': [
           {
             'address': 'nina@domain.com',
+          },
+          {
+            'address': 'nina1@domain.com',
+          },
+          {
+            'address': 'nina2@domain.com',
+          },
+          {
+            'address': 'nina3@domain.com',
           }
         ],
       }
@@ -124,60 +162,163 @@ void main() {
     });
 
     test(
-        'fetchContacts should return a list of contacts without duplicated phone number',
+        'Give a list of phonebook contacts'
+        'When fetchContacts is called'
+        'Then it should return a list of contacts without duplicated phone number',
         () async {
       final List<Contact> expectedListContact = [
-        const Contact(
+        Contact(
+          id: 'id_1',
           displayName: 'Alice',
-          phoneNumber: '(212)555-6789',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '(212)555-6789',
+            ),
+          },
+          emails: {
+            Email(
+              address: 'Alice@domain.com',
+            ),
+            Email(
+              address: 'Alice_1@domain.com',
+            ),
+          },
         ),
-        const Contact(
+        Contact(
+          id: 'id_2',
           displayName: 'Bob',
-          phoneNumber: '2124678190',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '2124678190',
+            ),
+          },
+          emails: {
+            Email(
+              address: 'bob@domain.com',
+            ),
+            Email(
+              address: 'bob2@domain.com',
+            ),
+          },
         ),
-        const Contact(
+        Contact(
+          id: 'id_3',
           displayName: 'Charlie',
-          phoneNumber: '212 555-6789',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '212 555-6789',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_4',
           displayName: 'David',
-          phoneNumber: '2124678190',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '2124678190',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_5',
           displayName: 'Eve',
-          phoneNumber: '+1.123.456.7890',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '+1.123.456.7890',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_6',
           displayName: 'Frank',
-          phoneNumber: '81234977890',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '81234977890',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_7',
           displayName: 'Grace',
-          phoneNumber: '+1 (800)-555-1234 ext. 123',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '+1 (800)-555-1234 ext. 123',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_8',
           displayName: 'Hank',
-          phoneNumber: '18005879106234',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '18005879106234',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_9',
           displayName: 'Ivy',
-          phoneNumber: '+1 (800)-555.1234',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '+1 (800)-555.1234',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_10',
           displayName: 'Karl',
-          phoneNumber: '18005873456',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '18005873456',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_11',
           displayName: 'Liam',
-          phoneNumber: '(212) 555-6789',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '(212) 555-6789',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_12',
           displayName: 'Mia',
-          phoneNumber: '2125556789',
+          phoneNumbers: {
+            PhoneNumber(
+              number: '2125556789',
+            ),
+          },
+          emails: const {},
         ),
-        const Contact(
+        Contact(
+          id: 'id_13',
           displayName: 'Nina',
-          email: 'nina@domain.com',
+          phoneNumbers: const {},
+          emails: {
+            Email(
+              address: 'nina@domain.com',
+            ),
+            Email(
+              address: 'nina1@domain.com',
+            ),
+            Email(
+              address: 'nina2@domain.com',
+            ),
+            Email(
+              address: 'nina3@domain.com',
+            ),
+          },
         ),
       ];
